requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
aws-cdk.assets==1.69.0
aws-cdk.aws-apigateway==1.69.0
aws-cdk.aws-applicationautoscaling==1.69.0
aws-cdk.aws-autoscaling==1.69.0
aws-cdk.aws-autoscaling-common==1.69.0
aws-cdk.aws-autoscaling-hooktargets==1.69.0
aws-cdk.aws-batch==1.69.0
aws-cdk.aws-certificatemanager==1.69.0
aws-cdk.aws-cloudformation==1.69.0
aws-cdk.aws-cloudfront==1.69.0
aws-cdk.aws-cloudwatch==1.69.0
aws-cdk.aws-codebuild==1.69.0
aws-cdk.aws-codecommit==1.69.0
aws-cdk.aws-codedeploy==1.69.0
aws-cdk.aws-codeguruprofiler==1.69.0
aws-cdk.aws-codepipeline==1.69.0
aws-cdk.aws-codepipeline-actions==1.69.0
aws-cdk.aws-cognito==1.69.0
aws-cdk.aws-ec2==1.69.0
aws-cdk.aws-ecr==1.69.0
aws-cdk.aws-ecr-assets==1.69.0
aws-cdk.aws-ecs==1.69.0
aws-cdk.aws-efs==1.69.0
aws-cdk.aws-elasticloadbalancing==1.69.0
aws-cdk.aws-elasticloadbalancingv2==1.69.0
aws-cdk.aws-events==1.69.0
aws-cdk.aws-events-targets==1.69.0
aws-cdk.aws-iam==1.69.0
aws-cdk.aws-kinesis==1.69.0
aws-cdk.aws-kms==1.69.0
aws-cdk.aws-lambda==1.69.0
aws-cdk.aws-logs==1.69.0
aws-cdk.aws-route53==1.69.0
aws-cdk.aws-route53-targets==1.69.0
aws-cdk.aws-s3==1.69.0
aws-cdk.aws-s3-assets==1.69.0
aws-cdk.aws-sam==1.69.0
aws-cdk.aws-secretsmanager==1.69.0
aws-cdk.aws-servicecatalog==1.69.0
aws-cdk.aws-servicediscovery==1.69.0
aws-cdk.aws-sns==1.69.0
aws-cdk.aws-sns-subscriptions==1.69.0
aws-cdk.aws-sqs==1.69.0
aws-cdk.aws-ssm==1.69.0
aws-cdk.aws-stepfunctions==1.69.0
aws-cdk.cloud-assembly-schema==1.69.0
aws-cdk.core==1.69.0
aws-cdk.custom-resources==1.69.0
aws-cdk.cx-api==1.69.0
aws-cdk.pipelines==1.69.0
aws-cdk.region-info==1.69.0
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
constructs==3.0.29
idna==2.10
iniconfig==1.1.1
jsii==1.13.0
packaging==20.4
pluggy==0.13.1
publication==0.0.3
py==1.9.0
pycryptodome==3.9.8
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.25.11
